Overhauled data virtualization
Greatly improved performance of loading items as they become visible in the scroll area. Items used to be realized individually. However, this generated a lot of background threads and resulted in jitters and stutters if the user scrolled through the list of items very quickly. Now, items are loaded in batches of 50 items per batch, which reduces the number of threads generated. Scrolling is much smoother now.

As a result, the implementation has changed greatly. ObservableVector delegates item retrieval to IBatchedItemProvider. This interface defines a pre-.NET 4 style async contract with events that get raised to indicate the completion of operations. PersonaViewModelProvider implements this interface, and has a mode to force synchronous execution. The reason for this is because the shitty XAML Designer can't load UIs from async code. Additionally, the PersonaViewModel uses BatchCalculator to determine which items to retrieve for a batch and uses a refactored IPersonaService instance to accomplish this.