Save and restore page state
Rolled out my own solution for saving page state on app suspension and restoration after termination. Specifically, I took a ViewModel First approach:
* SuspensionManager tracks navigating from/to a page and will use IAppState to save and restore state for an object that was registered for the page. For our purposes, a view model is registered with a page whose state should be tracked.
* IAppState is a generic interface that takes keys and values for caching state, and can persist and restore its key/value collection.
* JsonAppState implements IAppState and stores all values as JSON tokens. Note that if the value passed in implements IJsonSerializable then the value is cast and Write() is called to retrieve the JSON token.
* IJsonSerialiable is for any object that can write out and read in a JSON token.

So far, state restoration works if the app is terminated from the PersonaDetailsPage. Currently, if the app was terminated on the MainPage, and no Persona had been selected, the scroll offset isn't restored. This is because the MainPage wasn't navigated away from, so the SuspensionManager will not save the MainViewModel's state.